# docs/openapi/archon.yaml
openapi: 3.1.0
info:
  title: Archon
  version: standalone-trusted-v1
  description: Trusted-network coordinator and authoring API. JSON uses the success/timestamp/data envelope except for the raw theme document,
    with error details on failure. No authentication. Static cockpit serving is optional. Authoring request
    details follow internal/api.go and agents.go; run projections are sanitized, and run evidence routes serve
    capped, confined content (ADR-0017).
paths:
...
  /api/agents:
    get:
      responses:
        '200':
          description: Roster in data.agents and data.count, including presets and overrides, and
            data.harnesses, the harnesses whose seats Archon starts, each with the efforts it accepts and its
            defaultEffort, and data.effortPolicy, the {effort,use} lines that guide a slot's effort (low for
            errands, medium for making things, xhigh for architecture and review, max for consequential
            reviews). No live-session provider is wired by the daemon.
        '422':
          description: INVALID_AGENT_CARD when a stored card lacks a required field; the message names the card and field
      parameters:
      - in: query
        name: capable
        schema:
          type: string
      - in: query
        name: assignable
        schema:
          type: boolean
    post:
      responses:
        '201':
          description: Created persona card in data; returns ETag
        '422':
          description: INVALID_AGENT_CARD when the request gives a model or effort; a new role carries neither,
            and each slot that uses it sets them
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
  /api/agents/{agentId}:
    parameters:
    - in: path
      name: agentId
      required: true
      schema:
        type: string
    get:
      responses:
        '200':
          description: Persona card in data; returns ETag. Each harnessVariants entry has its stored model and
            effort, effectiveEffort (medium when effort is unset), efforts (the values its harness accepts), and
            seatLaunch, the command a seat for that variant runs, rendered by the seat launcher on the daemon's
            PATH; seatLaunchError says why a variant cannot start (an unsupported harness, or its CLI missing
            from PATH). Create and patch return the same.
        '422':
          description: INVALID_AGENT_CARD when the stored card lacks a required field; the message names the card and field
    patch:
      responses:
        '200':
          description: Updated persona card in data; returns ETag
        '422':
          description: INVALID_AGENT_CARD when model or effort is not one the harness accepts, or variant names
            no harness variant of the card
        '428':
          description: Missing If-Match
        '409':
          description: Stale ETag
      parameters:
      - in: header
        name: If-Match
        required: true
        schema:
          type: string
        description: Current ETag of the resource being edited
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentPatch'
components:
...
  schemas:
...
    AgentCreate:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        kind:
          type: string
        summary:
          type: string
        capabilities:
          type: array
          items:
            type: string
        personality:
          type: string
        harness:
          type: string
        sessionStem:
          type: string
        model:
          type: string
          description: Refused when non-empty. A new role carries no model; each slot sets its own.
        effort:
          type: string
          description: Refused when non-empty. A new role carries no effort; each slot sets its own.
        source:
          type: string
    AgentPatch:
      type: object
      properties:
        addCapability:
          type: string
        removeCapability:
          type: string
        addHarness:
          type: string
        sessionStem:
          type: string
        variant:
          type: string
          description: Harness variant that model and effort change; defaults to the card's default harness.
            With addHarness, model and effort set the added variant instead.
        variants:
          type: array
          description: Model and effort for several variants in one edit; all apply or none do. Name each
            variant once, and not also through variant/model/effort (422 otherwise). Only the fields given are
            validated.
          items:
            type: object
            required: [id]
            properties:
              id:
                type: string
              model:
                type: string
              effort:
                type: string
        model:
          type: string
          description: Model for the variant; an empty string clears it to the harness default model.
        effort:
          type: string
          description: Effort the variant's harness accepts; an empty string clears it to medium.
        source:
          type: string
        note:
          type: string
        retire:
          type: boolean
        displayName:
          type: string
        kind:
          type: string
        summary:
          type: string
        capabilities:
          type: array
          items:
            type: string
      description: Persona editor fields. Model and effort edit one harness variant (variant, else the
        default harness). Source is harness provenance, not replacement card content.